# chore: Update dependencies and benchmark relating code