--- a/sandbox/GeneratorSandbox/GeneratorSandbox.csproj
+++ b/sandbox/GeneratorSandbox/GeneratorSandbox.csproj
@@ -18,8 +18,8 @@
 
   <ItemGroup>
     <!--<PackageReference Include="Microsoft.Extensions.Configuration.Json" Version="9.0.0" />
-		<PackageReference Include="Microsoft.Extensions.Options.ConfigurationExtensions" Version="9.0.0" />-->
-    <PackageReference Include="Microsoft.Extensions.Hosting" Version="9.0.0" />
+    <PackageReference Include="Microsoft.Extensions.Options.ConfigurationExtensions" Version="9.0.0" />-->
+    <PackageReference Include="Microsoft.Extensions.Hosting" Version="9.0.6" />
     <!--<PackageReference Include="Microsoft.Extensions.DependencyInjection" Version="9.0.0" />-->
     <!--<PackageReference Include="ZLogger" Version="2.5.9" />-->
   </ItemGroup>