@filzrev 2025-06-20
I've compared to README benchmark results that is  measured on 2024/5/31.
But result is apparently different for `System.CommandLine`

https://private-user-images.githubusercontent.com/46207/335307159-db4bf599-9fe0-4ce4-801f-0003f44d5628.png?jwt=eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpc3MiOiJnaXRodWIuY29tIiwiYXVkIjoicmF3LmdpdGh1YnVzZXJjb250ZW50LmNvbSIsImtleSI6ImtleTUiLCJleHAiOjE3NTAzODM2NzgsIm5iZiI6MTc1MDM4MzM3OCwicGF0aCI6Ii80NjIwNy8zMzUzMDcxNTktZGI0YmY1OTktOWZlMC00Y2U0LTgwMWYtMDAwM2Y0NGQ1NjI4LnBuZz9YLUFtei1BbGdvcml0aG09QVdTNC1ITUFDLVNIQTI1NiZYLUFtei1DcmVkZW50aWFsPUFLSUFWQ09EWUxTQTUzUFFLNFpBJTJGMjAyNTA2MjAlMkZ1cy1lYXN0LTElMkZzMyUyRmF3czRfcmVxdWVzdCZYLUFtei1EYXRlPTIwMjUwNjIwVDAxMzYxOFomWC1BbXotRXhwaXJlcz0zMDAmWC1BbXotU2lnbmF0dXJlPTYyZTBhOTg4MzhiODY5ZjVmZTM1MjQ3Y2E2NjYyMTI3YjQyNDViOWVjNjdkYjFkMTI5MTAyOTk4MzVjM2RhYzMmWC1BbXotU2lnbmVkSGVhZGVycz1ob3N0In0.hBNX0-Srh305CJt39I62qfuDB6qttuRZMo9b1uCUs1E

I don't know what cause this performance differences.
I also tried `v5.0.0` version benchmarks but results are almost same.